+++ OneBlink.SDK/models/FormSubmissionEventTrimUriOption.cs
@@ -5,6 +5,6 @@ namespace OneBlink.SDK.Model
     public class FormSubmissionEventTrimUriOption
     {
         public string label { get; set; }
-        public int uri { get; set; }
+        public long uri { get; set; }
     }
 }
